refactor: 코드 리팩토링
- 변수명 변경
- checkWiningNumber 메소드를 간소화 -> checkNumberOfMap 코드 생성
- LottoGame의 createLottoBalls 메소드를 Lottos로 이동
