+++ src/main/java/domain/Lottos.java
@@ -5,13 +5,15 @@ import view.OutputView;
 import java.util.*;
 
 public class Lottos {
+    private static final int LOTTO_FIRST_NUMBER = 1;
+    private static final int LOTTO_LAST_NUMBER = 45;
     private static final int LOTTO_SIZE = 6;
     private static final int LOTTO_PRICE = 1000;
-    private static final int RANK_MINIMAL_NUMBER = 3;
 
     private final List<Lotto> lottos = new ArrayList<>();
 
-    public void createLotto(int purchasePrice, List<Integer> lottoBalls) {
+    public void createLotto(int purchasePrice) {
+        List<Integer> lottoBalls = createLottoBalls();
         int numberOfLottoTicket = purchasePrice / LOTTO_PRICE;
 
         System.out.println(numberOfLottoTicket + "개를 구매하셨습니다.");
@@ -28,25 +30,27 @@ public class Lottos {
     }
 
     public Map<Rank, Integer> checkWiningNumber(int[] winingNumber, int bonusNumber) {
-        // 로또번호가 일치한 개수와 배열 인덱스가 일치하도록 배열사이즈를 정함
-        //int[] numberOfWins = new int[LOTTO_SIZE + 1];
-        Map<Rank, Integer> numberOfWins = new EnumMap<>(Rank.class);
-        numberOfWins.put(Rank.FIFTH, 0);
-        numberOfWins.put(Rank.FOURTH, 0);
-        numberOfWins.put(Rank.THIRD, 0);
-        numberOfWins.put(Rank.SECOND, 0);
-        numberOfWins.put(Rank.FIRST, 0);
+        Map<Rank, Integer> numberOfWins = createNumberOfWinsMap();
 
         for (Lotto lotto : lottos) {
             int matchedCount = lotto.check(winingNumber);
-            Boolean isMatchBonusNumber = lotto.isMatchBonusNumber(bonusNumber);
+            boolean isMatchBonusNumber = lotto.isMatchBonusNumber(bonusNumber);
 
             Rank rank = Rank.checkRank(matchedCount, isMatchBonusNumber);
             if (numberOfWins.containsKey(rank)) {
                 numberOfWins.put(rank, numberOfWins.get(rank) + 1);
             }
         }
+        return numberOfWins;
+    }
 
+    private Map<Rank, Integer> createNumberOfWinsMap() {
+        Map<Rank, Integer> numberOfWins = new EnumMap<>(Rank.class);
+        numberOfWins.put(Rank.FIFTH, 0);
+        numberOfWins.put(Rank.FOURTH, 0);
+        numberOfWins.put(Rank.THIRD, 0);
+        numberOfWins.put(Rank.SECOND, 0);
+        numberOfWins.put(Rank.FIRST, 0);
         return numberOfWins;
     }
 
@@ -60,6 +64,16 @@ public class Lottos {
         }
 
         return ((benefit / money) - 1) * 100;
+    }
+
+    private List<Integer> createLottoBalls() {
+        List<Integer> lottoBalls = new ArrayList<>();
+
+        for (int lottoBall = LOTTO_FIRST_NUMBER; lottoBall <= LOTTO_LAST_NUMBER; lottoBall++) {
+            lottoBalls.add(lottoBall);
+        }
+
+        return lottoBalls;
     }
